Former-commit-id: 5856957c577a17bedee76c245fb6e60a75f56735 [formerly 4563e38db2d66490239ca51f30d0430c30fb2128 [formerly 3078b69273c308e0982ed6eb6fc76b0505ddc045] [formerly 5856957c577a17bedee76c245fb6e60a75f56735 [formerly 7820be8f078be179c7963cbae73cd3023a70c306]]]
Former-commit-id: 4563e38db2d66490239ca51f30d0430c30fb2128 [formerly 3078b69273c308e0982ed6eb6fc76b0505ddc045]
Former-commit-id: 4563e38db2d66490239ca51f30d0430c30fb2128
Former-commit-id: adb51300d8c44b7450b60d1d6dd14d45de6d0441

+++ cave/com.raytheon.viz.mpe.ui/src/com/raytheon/viz/mpe/ui/MPEDisplayManager.java
@@ -102,7 +102,8 @@ import com.raytheon.viz.ui.editor.IMultiPaneEditor;
  * Mar 14, 2013   1457     mpduff       Reset the gages on the resource.
  * Apr 18, 2013   1920     mpduff       Added updateGages method to reload the gage data, 
  *                                      fix formatting of legend for Base field Height.
- * Jul 02, 2013   2160     mpduff       Initialize newly displayed resources.                                     
+ * Jul 02, 2013   2160     mpduff       Initialize newly displayed resources.
+ * Feb 19, 2014   2628     mpduff       Change cast from short to int when creating color bar.
  * 
  * </pre>
  * 
@@ -750,7 +751,7 @@ public class MPEDisplayManager {
                     listener.displayFieldChanged(oldField, fieldToDisplay);
                 }
             }
-            }
+        }
 
         displayedFieldResource.issueRefresh();
     }
@@ -1020,13 +1021,12 @@ public class MPEDisplayManager {
                 entry.setOperator("<");
                 entry.setLabel("");
             } else {
-                // Convert display to data, cast to short, convert back to
-                // display
-                entry.setDisplayValue(dataToDisplay
-                        .convert((short) displayToData.convert(threshold)));
+                // Convert display to data, cast to int, convert back to display
+                entry.setDisplayValue(dataToDisplay.convert((int) displayToData
+                        .convert(threshold)));
                 if (displayField != DisplayFieldData.Index) {
                     entry.setLabel(format.format(threshold));
-            }
+                }
             }
             entry.setPixelValue((double) i);
